Rename tools/templating to tools/jinja2
