+++ pipelex/cogt/templating/template_blueprint.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel, Field, model_validator
 
 from pipelex.cogt.templating.template_category import TemplateCategory
 from pipelex.cogt.templating.templating_style import TemplatingStyle
-from pipelex.tools.templating.jinja2_parsing import check_jinja2_parsing
+from pipelex.tools.jinja2.jinja2_parsing import check_jinja2_parsing
 
 
 class TemplateBlueprint(BaseModel):
