+++ pipelex/cogt/templating/template_category.py
@@ -4,8 +4,8 @@ from typing import Any
 from jinja2.runtime import Context
 
 from pipelex.cogt.templating.templating_style import TextFormat
-from pipelex.tools.templating.jinja2_filters import tag, text_format
-from pipelex.tools.templating.jinja2_models import Jinja2FilterName
+from pipelex.tools.jinja2.jinja2_filters import tag, text_format
+from pipelex.tools.jinja2.jinja2_models import Jinja2FilterName
 from pipelex.types import StrEnum
