+++ pipelex/tools/jinja2/jinja2_filters.py
@@ -4,8 +4,8 @@ from jinja2 import pass_context
 from jinja2.runtime import Context, Undefined
 
 from pipelex.cogt.templating.templating_style import TagStyle, TextFormat
-from pipelex.tools.templating.jinja2_errors import Jinja2ContextError
-from pipelex.tools.templating.jinja2_models import Jinja2ContextKey, Jinja2TaggableAbstract
+from pipelex.tools.jinja2.jinja2_errors import Jinja2ContextError
+from pipelex.tools.jinja2.jinja2_models import Jinja2ContextKey, Jinja2TaggableAbstract
 from pipelex.types import StrEnum
 
 ########################################################################################
