+++ pipelex/tools/jinja2/jinja2_required_variables.py
@@ -5,8 +5,8 @@ from jinja2.exceptions import (
 )
 
 from pipelex.cogt.templating.template_category import TemplateCategory
-from pipelex.tools.templating.jinja2_environment import make_jinja2_env_without_loader
-from pipelex.tools.templating.jinja2_errors import Jinja2DetectVariablesError, Jinja2StuffError
+from pipelex.tools.jinja2.jinja2_environment import make_jinja2_env_without_loader
+from pipelex.tools.jinja2.jinja2_errors import Jinja2DetectVariablesError, Jinja2StuffError
 
 
 def detect_jinja2_required_variables(
