+++ tests/unit/pipelex/tools/templating/test_jinja2.py
@@ -5,7 +5,7 @@ import pytest
 from pipelex import log, pretty_print
 from pipelex.cogt.templating.template_category import TemplateCategory
 from pipelex.cogt.templating.templating_style import TagStyle, TemplatingStyle, TextFormat
-from pipelex.tools.templating.jinja2_rendering import render_jinja2
+from pipelex.tools.jinja2.jinja2_rendering import render_jinja2
 from tests.cases import Fruit, JINJA2TestCases
 
 PLACE_HOLDER = "place_holder"
